Adapt _DownloadOnlySubdirData to be compatible with newer conda SubdirData class (#119)
* Adapt _DownloadOnlySubdirData to be compatible with conda post
https://github.com/conda/conda/pull/12003

Functions `_process_raw_repodata_str` and `_process_raw_repodata` were
easy enough to make cross compatible. But `_read_local_repdata` does
not have an easy way to make compatible with both versions. The
function signatures are:

 - `old` `_read_local_repdata(self, etag, mod_stamp)`
 - `new` `_read_local_repdata(self, state):`

The compromise I settled on was to just use `def
_read_local_repdata(self, *args, **kwargs)` instead of adding
something that sort of matches the old and new interface.

* Update news/119-breaking-conda-internal-api

---------

Co-authored-by: jaimergp <jaimergp@users.noreply.github.com>