Upcoming changes in mamba context breaks conda-libmamba-solver (#193)
* Upcoming changes in mamba context breaks conda-libmamba-solver

The following commit to mamba https://github.com/mamba-org/mamba/pull/2432/ changes the way in which the context object is constructed. This means the next libmambapy version will break conda-libmamba-solver. Will be opening a PR on the fixes but the bigger question is how to preserve compatibility with older libmamba versions.

We are going to have to continue with repodata patches potentially due to this? Hope this issue/PR opens a larger discussion.

```python
ctx = libmambapy.context.Context()
ctx.json # Attribute Error
ctx.output_config.json # exists
```

This PR should not be merged since mamba with these changes has not
been released yet.

* add getattr safeguards ans group things together

* pre-commit

* add news [skip ci]

---------

Co-authored-by: jaimergp <jaimergp@users.noreply.github.com>